Refactor extract LoginDialog to 'gui.auth' namespace
