+++ securedrop_client/locale/messages.pot
@@ -59,42 +59,6 @@ msgstr ""
 msgid "Failed to delete source at server"
 msgstr ""
 
-msgid "Username"
-msgstr ""
-
-msgid "Passphrase"
-msgstr ""
-
-msgid "Two-Factor Code"
-msgstr ""
-
-msgid "SecureDrop Client v{}"
-msgstr ""
-
-msgid "SIGN IN"
-msgstr ""
-
-msgid ""
-"That username won't work.\n"
-"It should be at least 3 characters long."
-msgstr ""
-
-msgid ""
-"That passphrase won't work.\n"
-"It should be between 14 and 128 characters long."
-msgstr ""
-
-msgid ""
-"That two-factor code won't work.\n"
-"It should only contain numerals."
-msgstr ""
-
-msgid "SIGNING IN"
-msgstr ""
-
-msgid "Please enter a username, passphrase and two-factor code."
-msgstr ""
-
 msgid "SecureDrop Client {}"
 msgstr ""
 
@@ -111,6 +75,9 @@ msgid "{}"
 msgstr ""
 
 msgid "SIGN OUT"
+msgstr ""
+
+msgid "SIGN IN"
 msgstr ""
 
 msgid "Nothing to see just yet!"
@@ -233,6 +200,9 @@ msgstr ""
 msgid "Remember to be careful when working with files outside of your Workstation machine."
 msgstr ""
 
+msgid "Passphrase"
+msgstr ""
+
 msgid "SUBMIT"
 msgstr ""
 
@@ -308,6 +278,36 @@ msgid "Entire source account"
 msgstr ""
 
 msgid "Files and messages"
+msgstr ""
+
+msgid "Username"
+msgstr ""
+
+msgid "Two-Factor Code"
+msgstr ""
+
+msgid "SecureDrop Client v{}"
+msgstr ""
+
+msgid ""
+"That username won't work.\n"
+"It should be at least 3 characters long."
+msgstr ""
+
+msgid ""
+"That passphrase won't work.\n"
+"It should be between 14 and 128 characters long."
+msgstr ""
+
+msgid ""
+"That two-factor code won't work.\n"
+"It should only contain numerals."
+msgstr ""
+
+msgid "SIGNING IN"
+msgstr ""
+
+msgid "Please enter a username, passphrase and two-factor code."
 msgstr ""
 
 msgid "USE OFFLINE"
